cleanup

+++ rlclientlib/dedup.cc
@@ -269,7 +269,7 @@ struct dedup_collection_serializer
     return error_code::success;
   }
 
-  int finalize(api_status* status, unsigned long long int original_event_count) { return finalize(status); }
+  int finalize(api_status* status, uint64_t original_event_count) { return finalize(status); }
 
   int _dummy;
   shared_state_t& _state;
